+++ src/Quartz.Benchmark/RAMJobStoreBenchmark.cs
@@ -1,4 +1,4 @@
-using Quartz.Tests;
+﻿using Quartz.Tests;
 using BenchmarkDotNet.Attributes;
 using Quartz.Impl.Matchers;
 using Quartz.Impl.Triggers;