+++ src/Quartz/SPI/TriggerFiredBundle.cs
@@ -1,4 +1,4 @@
-#region License
+﻿#region License
 
 /*
  * All content copyright Marko Lahma, unless otherwise indicated. All rights reserved.
@@ -30,7 +30,6 @@ namespace Quartz.Spi;
 /// <seealso cref="QuartzScheduler" />
 /// <author>James House</author>
 /// <author>Marko Lahma (.NET)</author>
-[Serializable]
 public sealed class TriggerFiredBundle
 {
     /// <summary>